devShell: cleanups

+++ modules/devshell.nix
@@ -19,24 +19,31 @@ with builtins; let
             default = { };
             description = "Environment variables to set in the dev shell.";
           };
-          cleanupPaths = lib.mkOption {
+          cleanups = lib.mkOption {
             type = lib.types.lazyAttrsOf (lib.types.oneOf [
               lib.types.str
               (lib.types.submodule {
                 options = {
                   path = lib.mkOption {
-                    type = lib.types.str;
+                    type = lib.types.nullOr lib.types.str;
+                    default = null;
                     description = "Path to clean up.";
                   };
+                  script = lib.mkOption {
+                    type = lib.types.nullOr lib.types.str;
+                    default = null;
+                    description = "Custom cleanup script.";
+                  };
                   type = lib.mkOption {
-                    type = lib.types.enum [ "symlink" "file" "dir" ];
-                    description = "Type of item: symlink, file, or dir.";
+                    type = lib.types.enum [ "symlink" "file" "dir" "script" ];
+                    default = "script";
+                    description = "Type of item: symlink, file, dir, or script.";
                   };
                 };
               })
             ]);
             default = { };
-            description = "AttrSet of paths/lines or { path, type } objects to clean up in devShell.";
+            description = "AttrSet of paths/lines or { path|script, type } objects to clean up in devShell.";
           };
           overrides = lib.mkOption {
             type = lib.types.lazyAttrsOf (lib.types.anything);
@@ -46,39 +53,32 @@ with builtins; let
         };
       };
 
-      config = let
-        cleanupPathsList = attrValues config.myDevShell.cleanupPaths;
-        genCleanupCmd = item:
-          if isString item then ''
-            if [ -L "${item}" ]; then
-              unlink "${item}"
-            elif [ -f "${item}" ]; then
-              rm -f "${item}"
-            elif [ -d "${item}" ]; then
-              rm -rf "${item}"
-            fi
-          ''
-          else if item.type == "symlink" then ''
-            [ -L "${item.path}" ] && unlink "${item.path}"
-          ''
-          else if item.type == "file" then ''
-            [ -f "${item.path}" ] && rm -f "${item.path}"
-          ''
-          else if item.type == "dir" then ''
-            [ -d "${item.path}" ] && rm -rf "${item.path}"
-          ''
-          else "";
-        cleanupScript = pkgs.writeShellScriptBin "devshell-cleanup" (
-          lib.concatMapStringsSep "\n" genCleanupCmd cleanupPathsList
-        );
-      in {
-        devShells.default = (pkgs.mkShell.override config.myDevShell.overrides {
-          # env = config.myDevShell.env;
-          env = lib.mapAttrs (_: v: if (isBool v) then (if v then "true" else "false") else v) config.myDevShell.env;
-          buildInputs = config.myDevShell.buildInputs ++ [ cleanupScript ];
-          shellHook = lib.concatStringsSep "\n" (attrValues config.myDevShell.shellHooks);
-        });
-      };
+      config =
+        let
+          cleanupsList = attrValues config.myDevShell.cleanups;
+          genCleanupCmd = item:
+            if isString item then ''
+              if [ -L "${item}" ]; then unlink "${item}"
+              elif [ -f "${item}" ]; then rm -f "${item}"
+              elif [ -d "${item}" ]; then rm -rf "${item}"
+              fi
+            ''
+            else if item.type == "script" || item.script != null then item.script
+            else if item.type == "symlink" then '' [ -L "${item.path}" ] && unlink "${item.path}" ''
+            else if item.type == "file" then '' [ -f "${item.path}" ] && rm -f "${item.path}" ''
+            else if item.type == "dir" then '' [ -d "${item.path}" ] && rm -rf "${item.path}" ''
+            else "";
+          cleanupScript = pkgs.writeShellScriptBin "devshell-cleanup" (
+            lib.concatMapStringsSep "\n" genCleanupCmd cleanupsList
+          );
+        in
+        {
+          devShells.default = (pkgs.mkShell.override config.myDevShell.overrides {
+            env = lib.mapAttrs (_: v: if (isBool v) then (if v then "true" else "false") else v) config.myDevShell.env;
+            buildInputs = config.myDevShell.buildInputs ++ [ cleanupScript ];
+            shellHook = lib.concatStringsSep "\n" (attrValues config.myDevShell.shellHooks);
+          });
+        };
     };
   };
